fix: stop nvim-cmp from destroying update autocmd
Issue happens since 'FileType' was added to MulticolumnReload (bf232e9).

I'm... not entirely sure how to best handle this issue. I don't yet
understand *why* the current buffer and window become the cmp_* stuff.
Leaving this ugly hack for now, because the only alternative I can think
of is to stop reloading on 'FileType', which is bad in a different way.