fix: colorcolumn disappearing
There's a bit to unpack on this one:

Use only WinEnter event for reload() and change how the function works
to more reliably detect when buffer/window is disabled, which was done
by delaying some of the logic in a deferred call.

Furthermore, this commits refactors some code in the column module to be
more window ID oriented, which seems to makes more sense (highlights are
per window, colorcolumns aren't rendered to not visible buffers, etc.).